+++ test-support/assertions/is-false.js
@@ -1,5 +1,7 @@
+import assertionMessage from '../helpers/assertion-message';
+
 export default function isFalse(context, testee, message = "") {
   const result       = testee === false;
-  const finalMessage = (message.length ? message + ": " : '') + "Expected to be false.";
+  const finalMessage = assertionMessage("Expected to be false.", message);
   this.push(result, testee, false, finalMessage);
 }