+++ test-support/assertions/numbers-almost-equal.js
@@ -1,11 +1,19 @@
 import _ from 'npm:lodash';
+import assertionMessage from '../helpers/assertion-message';
 
-export default function numbersAlmostEqual(context, actual, expected, message = "") {
+export default function numbersAlmostEqual(context, actual, expected, precision, message = "") {
 
-  // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/EPSILON#Testing_equality
-  const result = Math.abs(actual - expected) < Number.EPSILON;
+  if (_.isString(precision)) {
+    message = precision;
+  }
 
-  const finalMessage = `${message.length ? `${message}: ` : ''}Expected to be almost equal.`;
+  if (!_.isNumber(precision)) {
+    precision = 6;
+  }
+
+  const result       = Math.abs(actual - expected) <= (1 /  Math.pow(10, precision));
+  const finalMessage = assertionMessage('Expected to be almost equal.', message);
 
   this.push(result, actual, expected, finalMessage);
+
 }