+++ test-support/helpers/assertion-message.js
@@ -1,0 +1,3 @@
+export default function assertionMessage (foo, bar) {
+  return bar && bar.length ? `${bar}: ${foo}` : foo;
+}